rename folder for api request functions
