+++ apps/newsletters-ui/src/app/api-requests/requestDraftDeletion.ts
